Add the (massive) pyproject.toml!
This adds the pyproject.toml file, which contains all of the information
that would normally be included in `setup.cfg` or `setup.py`!

We are still using Setuptools for installation, so the build
requirements include very-new versions of setuptools and wheel.

At this time, three options are supported.

* `build` brings in pypa-build, for building source distributions.  It's
  just a convenience, and is not used for any code execution.

* `docs` brings in the tooling needed to build docs.

* `gcs` brings in the code needed to use Google Cloud Secret Manager,
  and support `gcs://` URLs in configuration variables.

There is one annoyance with this setup, which is that the version number
is hard-coded in the file.  I need to figure out something to let that
verion number propagate into `__version__`.