workflow: Push built docs to gh-pages automatically
This checks out the GitHub Pages branch; wipes it; copies the build
output to it; and runs a third-party action to add, commit, and push.

The action uses a generic "GitHub Pages" identity as the committer.  I
would like to add the identity of the person who made the code change,
but I don't see how I can pull that information easily.

The artifact upload has been stopped, since it's no longer needed.