Add configuration-loading code
This is the code responsible for pulling configuration from places.

Configuration is stored either in environment variables, or in a .env
file.  It supports indirection: If the environment variable's value is a
recognized URL, the actual value will be fetched from the URL.
Supported URLs include `file:` and `gcs:` for Google Cloud Secret
Manager secrets (if the `gcs` option is installed).