Add Globus Groups code
This is the code which interacts with Globus Groups.  It handles
creating Groups, deleting Groups, fetching members, and adding members.

Still to be done is code for removing members.

The code all works off of Globus Identity Usernames.  It handles
conversion to/from Globus Identity IDs.

One thing worth noting: The only time admins can be set through this
code is during Group creation.  Since the Workgroup is the source of
truth, the only people who should be added & removed through this
interface are members.