descope_usernames: Require domain, and use rsplit
This is all because Globus supports Google Login.

I switched to rsplit instead of split, and set a max of 1, because
logins through Google have Identity usernames in the form of
"<email_address>@accounts.google.com".  That means they'll have two `@`
characters.  Using rsplit and setting a max of 1 ensures the proper
domain (in this case, "accounts.google.com" is split off).