Make GlobusClients a base class; add GlobusServerClients
It is likely that we'll be moving to have a separate set of Globus
Client instances for end-user accesses, vs. for server accesses.  For
example, let's say a client is accessing a web app and has authenticated
through Globus' OAuth2 IdP.  In that case, we will have Globus Auth and
Globus Groups clients, but with different permissions compared to if we
authenticated as the confidential client.

To enforce this different context, GlobusClients is being turned into a
base class, with GlobusServerClients as a subclass.  At some point in
the future, a new subclass will be created to represent Globus User
Clients (or something like that).