Add checking of errors to Globus Group batch calls
The Globus Groups API calls for batch membership operations, when it has
failures on individual operations, it does not completely fail.
Instead, it returns a list of Identity IDs that were successfully added
(or removed), plus an optional list of Identity IDs that had errors.

This updates the add and remove code to first check for the # of entries
returned: # added/removed plus # of errors must equal the # of entries
in the batch.

Next, the erorrs themselves are checked.  If we got an error because a
person was already added (or removed), that's OK, and is ignored.  If we
get some other error, it is logged as a warning, but we do not raise an
exception.  (That could change in the future.)