Add support for listing Globus Groups
Given a Globus client, return the list of groups where the client is a
member.

This is designed to work with any GlobusClients object.  It's meant to
list memberships for the Confidential Client, but should be able to work
for user clients.

It doesn't collect all the information that Globus Groups stores, it
just collects the information we care about.