Use simpler Globus Auth client for server
I don't know why I missed this.  The SDK docs say that the
`ConfidentialAppAuthClient` does exactly what we need for our auth
client on the server side: It uses Basic authentication for all of its
calls out to Globus Auth, and it also provides some additional OAuth2.0
methods that we can use, which will be important later.

So, switch the Server Auth client.