Add Globus Clients for an end user
This adds a new subclass of `GlobusClients`, for authenticating as an
end user.

The `GlobusUserClients` class is meant to be instantiated at the end of
an OAuth 2.0 authorization code flow.  Included are two static methods
to create a Flow Manager and get a URL, plus a class method to exchange
a code for tokens.

The class method instantiates Globus Auth and Groups clients, the former
of which is pretty limited, since we only request the `openid` and
`profile` scopes.  We also record the user's ID, username, ID Provider
information, the SGGM access token, and the expiration datetime of the
token.

Space is provided to hold a refresh token, and the static methods
support it, but our instantiating class method does not: The Auth and
Groups clients only get access tokens, even if a refresh token is being
requested.  All three tokens expire at the same time, so what should
happen is, when we refresh the SGGM token, we should use a Dependent
Token Grant to get new Auth and Groups tokens.  The Globus Auth client
accessed via the GlobusServerClients instance does support that.  What's
missing is the code to automatically trigger such a thing on token
refresh.

Another thing missing is a way to properly capture the auth code when
Globus Auth redirects the user.  As we're not a native client, we're not
allowed to use Globus Auth's helper web page, so we would have to
implement our own.  Right now I don't see this as a big issue, because
the only place it's being used is in the plumbing CLI (which shouldn't
be used normally).  Eventually it'll be used in web pages, which is
totally fine.