ggm.globus.client: Add user logout method
Adds a method to `GlobusUserClients`, logging out a user.  All of the
tokens (GGM access, GGM refresh, Auth, and Groups) are revoked, using
the server's Auth client.  After that, the expiration time is set to
now, and most other attributes are replaced with values that should be
easy to recognize if they end up in logs.

The tokens themselves are not erased, as they are already useless.