ggm.globus.client: Add serialization
Updates `GlobusUserClients` to add methods which convert between an
instance and a dict, which is then suitable for further serialization.
The intent is for the new methods to be used with JSON Encoder and
Decoder classes.  This code is not used to pickle or un-pickle.

The new type `GlobusUserClientsDict` defines what dict is returned, and
the type of each member.  The serialized form (the dict) only contains
values that are basic Python types (so, UUIDs and datetimes are strings
and ints).  It is the caller's responsibility to check that a dict is
actually a `GlobusUserClientsDict` before calling to deserialize.