webapp: Start login & logout
Start handling login and logout.  Logging in works via /login to
redirect to Globus Auth, and /login/complete to handle the response
(which ultimately gets saved to the session).  A random state value is
stored in the session, and checked against the state returned to us when
/login/complete is called.  This check is important, as it's one part of
CSRF prevention.

NOTE: When testing locally, be sure to set the `SERVER_NAME` environment
variable to `localhost`, and to use `localhost` in the URL (instead of
127.0.0.1).  The redirect URI must be an approved one (that is, listed
in the Globus Auth client's list of allowed redirect URIs).  It is also
checked by Globus Auth when we exchange for tokens (another bit of CSRF
prevention).

Login supports providing a post-login redirect URL via session.

Logout works via /logout to invalidate tokens and redirect back to the
home page.