ggm.globus.client: Request & check domain
Since we know the domain we want our Globus Identities to be in, we can
check that when the authentication completes.  Not only that, we can use
a relatively new login URL parameter (session_required_single_domain) to
ask that the user log in with an identity from that domain.

The latter check is not perfect, though: It's part of the login URL
given to the client (via a redirect), so it can be manipulated by the
user.  Therefore, we still have to check the domain at the end of the
flow.