ggm.globus.clients: Require MFA on user login
This asks Globus Auth to require that an MFA device be used during
login.  This requires that the domain's IdP do multi-factor auth in the
first place, and that it indicates this to Globus.  With SAML auth
(including auth via CILogon), this can be asserted by including the
REFEDS MFA Profile (https://refeds.org/profile/mfa) string as an
AuthnContextClassRef.

This is also not perfect, because it's a parameter passed via the login
URL, and can be intercepted (and stripped out) by the end user.