ggm.globus.client: Update logout doc and check before logout
The logout function didn't have a doc, so that's been added.

Also, I now check at the start of logout to see if the User ID has been
set to the null UUID.  If it is the null UUID, then the user has already
been logged out, so we don't need to do anything.