ggm.globus.client: Add last_checked to user client
This adds a new field, `last_checked`, to `GlobusUserClients`.  It will
be used to store the datetime when the tokens were last checked for
validity.

We have to balance how often we call out to Globus for token validation.
If we call out to Globus on every page access, that will be a lot of
calls out to Globus.  If we don't check at all, and only check our
locally-stored expiration datetime, we won't be able to catch if our
tokens are invalidated early (which could happen if a user revokes
consent, for example).  So, to implement a middle ground, add a datetime
recording when the tokens were last checked for validity.

This commit adds the field, sets it to 'now' on instantiation, and
includes it when converting to/from dict.  What's next is to actually
check and update the field.