ggm.globus.client: Add is-logged-in check
This adds a check to `GlobusUserClients`, checking if the user is logged
in.  This does a number of checks that don't involve calling out to
Globus.  It checks if the instance's User ID is the Null UUID, if the
token has expired, and if the token will expire soon.  The latter check
is so that we don't get ourselves stuck with the token expiring in a
long-running operation.

This does not yet call out to Globus.  That'll come shortly!