webapp: Have Flask log everything
We are using Flask's logger for Flask messages, but that does not log
everything.

This change take's Flask's default handler, and moves it to be the
handler for the root.  That means it picks up logging for everything.
This can be good, especially if the WSGI server running Flask has their
own logging handler; Flask is able to be told that, and adjusts its
default handler to suit.

This change means we now get log entries from ggm, which is good.  We
though however have to raise the log level for some of our libraries, so
they don't write out tons of logs we don't want.