webapp: Make the logged-in check globus pre-request
Instead of doing the check when we want to, have the "Are we logged in?"
check take place automatically at the start of each request.  If the
user is logged in, then the GlobusUserClients instance is added to
`g.clients`.  If the user is not logged in, then `g.clients` is set to
`None`.