webapp: Actually handle login_url in /login
Although /login/complete was looking for a 'login_url' in the session
(which would contain a URL to redirect the user after successful login),
we didn't populate that session field in /login.  This does that.  We
also make sure the URL is quoted and un-quoted (it might not be
necessary, though).