ggm.environ: Add token-check grace period
Add `APP_TOKEN_CHECK_GRACE` to the list of static config options.  When
asked to check if a token is valid, skip the check if the token has been
checked in the last `APP_TOKEN_CHECK_GRACE` seconds.

This is used to limit the amount of times we reach out to Globus Auth
for token verification.  We can only validate tokens individually, and
it's possible to have one token invalidated while the others remain
valid, so when we decide to check, we need to check all of them.  That
can add lag to each request, so only check tokens once in a while.

This is hard-coded to 10 minutes (600 seconds).