ggm.globus.client: Add token-checking
This adds checking tokens with Globus Auth.  It's marked as an internal
instance method of `GlobusUserClientsDict`, and is called by
`is_logged_in` after it completes its other checks.

First there is a simple check for a Null Globus User ID.  If the token
has already been checked recently, then we assume it's OK, unless the
check is forced.

Assuming we got this far, we check the refresh token (if present), and
then the access tokens.  Although refresh logic hasn't been coded yet,
we make a provision for it: If a refresh token is present & good, an
invalid access code will trigger a refresh instead of a failure.

If there is no refresh token, any bad access token triggers a failure.

NOTE: If this fails, the caller should trigger a full logout, to ensure
everything is properly purged.

Assuming everything is OK, the last-checked time is updated to now.

NOTE: The last-check modification, and a possible refresh, will cause
the instance to mutate.  Be sure the mutated session gets stored on the
client!