ggm.workgroup: Add code to handle UUID conversion
Globus Groups are identified by UUIDs.  When written out, UUIDs are
pretty long.  So, let's have a way to convert from a UUID to a
relatively-compact string that is compatible as a Workgroup name.

This uses Base32 encoding, with the padding '=' characters removed and a
'g' prefix added (to ensure the name starts with a letter).